% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Mcmc-Dsp.R
\name{dsp}
\alias{dsp}
\title{MCMC sampler for day-specific probabilities of conception methodology}
\usage{
dsp(dspDat, nSamp = 10000, nBurn = 0, nThin = 1, hypGam = NULL,
  tuningGam = NULL, hypPhi = NULL, tuningPhi = 0.3,
  trackProg = "percent", progQuants = seq(0.1, 1, 0.1),
  saveToFile = FALSE, outPath = NULL)
}
\arguments{
\item{dspDat}{An object of \code{class} \code{\link{dspDat}}.}

\item{nSamp}{The number of post-burn-in scans for which to perform the
  sampler.}

\item{nBurn}{Number of sampler scans included in the burn-in phase.}

\item{nThin}{Value such that during the post-burn-in phase, only every
  \code{nThin}-th scan is recorded for use in posterior inference.  Default
  of \code{1} corresponds to every scan being retained.}

\item{hypGam}{Either \code{NULL} or a \code{list} containing hyperparameters
  to be specified for the exponentiated model regression coefficients.  None,
  some, or all of the hyperparameters can be/need be specified.

  Each exponentiated regression coefficient has a prior defined in terms of 5
  hyperparameters. These hyperparameters are the (i) prior probability of the
  point mass state, the (ii) shape and (iii) rate of the gamma distribution
  state, and the (iv) lower (v) and upper bounds of the gamma distribution
  state.

  If not specified by the function input, then a default value is provided
  for each of the hyperparameters.  These default parameters, correponding to
  their description in the preceeding paragraph, are (i) \code{0.5}, (ii)
  \code{1}, (iii) \code{1}, (iv) \code{0}, and (v) \code{Inf}.

  Exponentiated regression coefficient hyperparameter specifications must be
  provided as follows.  If the input to \code{hypGam} is \code{NULL}, then
  every hyperparameter is taken to be the default value.  If some of the
  hyperparameters are to be specified, then \code{hypGam} must be a
  \code{list} containing a sub-hierarchy of \code{list}s; each of these
  second-level \code{list}s must have the name of one of model design matrix
  variables.  Thus if non-\code{NULL}, then \code{hypGam} is a \code{list}
  containing between \code{1} and \code{q} \code{list}s, where \code{q} is
  the number of covariates in the model (after recoding categorical variables
  to dummy-variable form).

  Each second-level \code{list} in \code{hypGam} must contain between
  \code{1} and \code{5} \code{numeric} values with possible names (i)
  \code{p}, (ii) \code{a}, (iii) \code{b}, (iv) \code{bndL}, or (v)
  \code{bndU} corresponding to the hyperparameter description from before
  with matching index.  The order of the objects in either level of
  \code{hypGam} does not matter.}

\item{tuningGam}{Either \code{NULL} or a list containing one or more
  \code{numeric} objects each with the name of one of the model design matrix
  variables; the values are used as tuning parameters for the Metropolis step
  for regression coeffients corresponding to continuous predictor variables.
  Categorical variables do not have a Metropolis step and values provided for
  them are ignored.  If \code{NULL}, then default values are provided for
  every regression coefficient (that corresponds to a continuous predictor
  variable).  If some but not all of the tuning parameters for regression
  coefficients corresponding to continuous predictor variables are provided,
  then default values are provided for the remaining. The default tuning
  value for any regression coefficient is \code{0.25}.}

\item{hypPhi}{Either \code{NULL} or a \code{list} containing one or two
  \code{numeric} objects with names \code{c1} and/or \code{c2}.  If supplied,
  then these values correspond (respectively) to the shape and rate
  parameters of the prior (gamma) distribution for the variance parameter of
  the woman-specific fecundability multipliers.  If \code{NULL}, then default
  values are provided for both \code{c1} and \code{c2}, and if exactly one of
  either \code{c1} or \code{c2} are provided, then a default value is
  provided for the other.  The default values for \code{c1} and \code{c2} are
  \code{1} and \code{1}, respectively.}

\item{tuningPhi}{Metropolis tuning parameter for the variance parameter of
  the woman-specific fecundability mulitpliers. The proposal value for this
  variance parameter is sampled from a uniform distribution with support as
  determined by the tuning parameter.}

\item{trackProg}{One of either \code{"none"}, \code{"percent"}, or
  \code{"verbose"}; partial matching is supported.}

\item{progQuants}{Vector with values in (0,1].  Ignored if \code{trackProg}
  is specified as \code{"none"}.  If \code{trackProg} is one of
  \code{"percent"} or \code{"verbose"} then the specified output is printed
  whenever the post-burn-in progress of the sampler reaches one of the
  quantiles specified by \code{progQuants}.}

\item{saveToFile}{\code{logical} specifying whether the samples from the
  post-burn-in phase are to be either written to file or returned as
  \code{data.frame} objects.  Note that in either case output characterizing
  the model is returned by the sampler.}

\item{outPath}{String specifying the local path into which output files
  containing the MCMC samples are to be placed.  Ignored if \code{saveToFile}
  is \code{FALSE}.}
}
\value{
\code{dsp} returns a list containing the following objects

  \describe{ \item{\code{formula}}{Model \code{formula}, as passed to the
  \code{dsp} sampler through the input to the \code{dspDat} parameter.}

  \item{\code{hypGam}}{\code{list} containing a sub-hierarchy of
  \code{list}s, each containing the hyperparameter values used for the
  sampler for the regression coefficients.}

  \item{\code{tuningGam}}{********}

  \item{\code{hypPhi}}{Hyperparameters for the variance parameter of the
  woman-specific fecundability multipliers.}

  \item{\code{tuningPhi}}{Metropolis tuning parameter used for sampling the
  variance parameter of the woman-specific fecundability multipliers.}

  \item{\code{nSamp}}{Input to \code{nSamp} parameter.}

  \item{\code{nBurn}}{Input to \code{nBurn} parameter.}

  \item{\code{nThin}}{Input to \code{nThin} parameter.}

  \item{\code{outPath}}{If \code{saveToFile} specified as \code{TRUE}, then
  the input to \code{outPath} parameter.}

  \item{\code{phi}}{If \code{saveToFile} specified as \code{FALSE}, then a
  vector containing the post-burn-in samples for the variance parameter of
  the woman-specific fecundability multipliers.}

  \item{\code{xi}}{If \code{saveToFile} specified as \code{FALSE}, then a
  \code{data.frame} containing the post-burn-in samples for the
  woman-specific fecundability multipliers.}

  \item{\code{gam}}{If \code{saveToFile} specified as \code{FALSE}, then a
  \code{data.frame} containing the post-burn-in samples for the regression
  coefficients.}

  }
}
\description{
\code{dsp} is an MCMC sampler for the methodology proposed by Dunson and
Stanford in \emph{Bayesian Inferences on Predictors of Conception
Probabilities} (2005).
}
\details{
Takes preprocessed fertility data in the form of a
  \code{\link{dspDat}} object and performs an MCMC sampling algorithm for the
  Dunson and Stanford day-specific probabilities of conception methodology.

  Selection of the covariates to include in the model is performed when
  creating the \code{dspDat} object.
}
\author{
David A. Pritchard and Sam Berchuck, 2015
}
\references{
Dunson, David B., and Joseph B. Stanford. "Bayesian inferences on
  predictors of conception probabilities." \emph{Biometrics} 61.1 (2005):
  126-133.
}

